Refactor to use UUID to identify players
Bumped major version to indicate breaking changes.
More breaking changes likely, ideally with a migrator.

Note that plots themselves are still named based on the player's
name at the time, and the name value in the config should stay as the
player's original name to avoid breaking the referencing of plots.
Additionally, friends are still stored as names. Migrating that to
UUIDs will be a little more difficult, but will be done at some point.

Will include Player->UUID migrator later (current local one is janky)